requirements.txt
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
flask>=2.3.0
werkzeug>=2.3.0
ultralytics>=8.0.0
paddleocr>=2.7.0
opencv-python>=4.8.0
numpy>=1.24.0

